Исправлены все неточности и выполнены некоторые пожелания после тестирования релиз-кандидата

+++ system/controllers/wall/actions/delete.php
@@ -6,31 +6,48 @@ class actionWallDelete extends cmsAction {
 
         if (!$this->request->isAjax()){ cmsCore::error404(); }
 
-        $entry_id = $this->request->get('id');
+        $entry_id = $this->request->get('id', 0);
 
-        // Проверяем валидность
-        $is_valid = is_numeric($entry_id);
-
-        if (!$is_valid){
-            $result = array(
-                'error' => true,
+        if (!$entry_id){
+            return $this->cms_template->renderJSON(array(
+                'error'   => true,
                 'message' => LANG_ERROR
-            );
-            cmsTemplate::getInstance()->renderJSON($result);
+            ));
         }
 
-        $user = cmsUser::getInstance();
-
         $entry = $this->model->getEntry($entry_id);
+        if (!$entry){
+            return $this->cms_template->renderJSON(array(
+                'error'   => true,
+                'message' => LANG_ERROR
+            ));
+        }
+
+        $entry = cmsEventsManager::hook('wall_before_delete', $entry);
+
+        $controller = cmsCore::getController($entry['controller']);
+
+        // Получаем права доступа
+        $permissions = $controller->runHook('wall_permissions', array(
+            'profile_type' => $entry['profile_type'],
+            'profile_id'   => $entry['profile_id']
+        ));
+
+        if (!$permissions || !is_array($permissions) || !$permissions['delete']){
+            return $this->cms_template->renderJSON(array(
+                'error'   => true,
+                'message' => LANG_ERROR
+            ));
+        }
 
         $this->model->deleteEntry($entry_id);
 
-        $result = array(
+        cmsEventsManager::hook('wall_after_delete', $entry);
+
+        $this->cms_template->renderJSON(array(
             'error' => false,
             'message' => LANG_WALL_ENTRY_DELETED
-        );
-
-        cmsTemplate::getInstance()->renderJSON($result);
+        ));
 
     }
